Remove items from hdict in bcf_hdr_remove()
As bcf_hdr_remove() deletes hrec structs it needs to ensure that
the pointers to them in the bcf_hdr_aux_t::gen dictionary are
removed as well, otherwise callers to bcf_hdr_get_hrec() may get
a stale pointer to the deleted item.

Unfortunately bcf_hdr_remove_from_hdict() needs to allocate
some memory to find the items in the dictionary.  If that fails
it falls back to a search through the dictionary values to
find the item, so we can be sure that it will always succeed.

Enhance tests to ensure bcf_hdr_get_hrec() returns NULL for
removed records.