Store ref counted bcf_hdr_t pointer in BGZF private data
For bcf files, the header pointer hasn't always been passed
into bcf_read(), especially when using iterators.  As having
it available would be useful for VCF 4.4+ support, this works
around its absence by attaching a pointer to the header in
BGZF private data, which was previously unused for vcf/bcf.
It also adds reference counting to the header so that it can
be cleaned up safely irrespective of whether hts_close() or
bcf_hdr_destroy() was called first.  To avoid ABI breakage,
the reference count is stored in the bcf_hdr_aux_t struct.