Speed up updatephasing
By noting that we're only interested in the least-significant
bit of each GT value, it's possible to reduce the number of
branches in this function by doing bit manipulations on the
first byte of each stored value.  The common haploid and diploid
cases are also specialised so the inner loop on ploidy can
be avoided for those cases.